+++ ViewModel/StoreInfoUsps.php
@@ -32,18 +32,6 @@ class StoreInfoUsps implements ArgumentInterface
         return $result;
     }
 
-    private function flattenArray(array $array): array
-    {
-        $flatArray = array_map(function($item) {
-            if (!is_array($item) || !array_key_exists('content', $item)) {
-                throw new \InvalidArgumentException('Each item in the array should be an array with a "content" key.');
-            }
-            return $item['content'];
-        }, $array);
-
-        return array_values($flatArray);
-    }
-
     public function getStoreUsps(string $attribute): array
     {
         $path = sprintf('siteation_storeinfo_usps/%s/usps', $attribute);
@@ -58,37 +46,31 @@ class StoreInfoUsps implements ArgumentInterface
 
     public function getHeaderUsps(): array
     {
-        $usps = $this->getStoreUsps('header');
-        return !empty($usps) ? $this->flattenArray($usps) : [];
+        return $this->getStoreUsps('header');
     }
 
     public function getFooterUsps()
     {
-        $usps = $this->getStoreUsps('footer');
-        return !empty($usps) ? $this->flattenArray($usps) : [];
+        return $this->getStoreUsps('footer');
     }
 
     public function getCategoryUsps(): array
     {
-        $usps = $this->getStoreUsps('category');
-        return !empty($usps) ? $this->flattenArray($usps) : [];
+        return $this->getStoreUsps('category');
     }
 
     public function getProductUsps(): array
     {
-        $usps = $this->getStoreUsps('product');
-        return !empty($usps) ? $this->flattenArray($usps) : [];
+        return $this->getStoreUsps('product');
     }
 
     public function getCustom1Usps(): array
     {
-        $usps = $this->getStoreUsps('custom_1');
-        return !empty($usps) ? $this->flattenArray($usps) : [];
+        return $this->getStoreUsps('custom_1');
     }
 
     public function getCustom2Usps(): array
     {
-        $usps = $this->getStoreUsps('custom_2');
-        return !empty($usps) ? $this->flattenArray($usps) : [];
+        return $this->getStoreUsps('custom_2');
     }
 }