+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/Siteation/magento2-storeinfo/compare/1.0.0...main
+[Unreleased]: https://github.com/Siteation/magento2-storeinfo-usps/compare/1.0.0...main
+
+### Added
+- Option to pick a custom icon
 
 ## 1.1.0 - 2024-12-29
 ### Added